fix: only set the layout to a template if _layoutSetup is passd; fix and consolidate tests
