+++ packages/common/src/core/_internal/getTemplate.ts
@@ -208,7 +208,7 @@ export function getTemplate(
 
   const blankSiteOrPageLayout: IHubLayout = {
     sections: [],
-    header: undefined,
+    header: { component: {} },
     footer: undefined,
   };
